+++ docs/changelog.md
@@ -8,13 +8,14 @@
         * `pyrotech.effect.comfort`
 
   * **Added:**
-    * Comfort effect
+    * Comfort effect:
         * Regenerates health exactly like the Regeneration effect
         * Increases hunger and saturation restored from eating
         * Added config values:
             * `COMFORT_HUNGER_MODIFIER`
             * `COMFORT_SATURATION_MODIFIER`
             * `COMFORT_REGEN_INTERVAL_TICKS`
+            * `COMFORT_REGEN_HALF_HEARTS`
     * Campfire will now add comfort effect at night if player is within range
         * Only works at night, by default between 13000 and 23000 or 19:00 and 5:00
         * Range depends on light level with a maximum range of 6 blocks and a minimum range of 1 block